This example now works for full rank textbook example matrix.
Still having some issues getting my parity check matrices into full rank form, or getting the encoding algorithm to handle matrices that are not full rank. I will figure out one of those two paths, but for now, here is an example of the encoding process working for a sample matrix from a textbook that is full rank.