Correcting some fundamental issues with getSystematicGmatrix.
Requiring that a matrix going through this algorithm be full rank. I had coded it such that a matrix of non-full-rank would be handled by identifying the dependent (non-indpendent) rows and then later, delete those row numbers from the orignal H matrix. This doesn't work because the column weights will be altered and the parity check H matrix is no longer regular. I am not ready to handle irregular H matrices yet, so making this a restriction on this function for now.