Trying to get a nonsingular phi matrix out of greedyUpperTriangulation function.
The authors of Modern Coding Theory suggest doing some row swappng to get a nonsingular phi, so I have added some steps to do that. It is still not working. One more thing I could draw is shuffling the rows of the [E C D] submatrix of H, but I think the real problem is that this algorithm will just not work with a parity check H matrix that is not full rank.