Adding initial constraint on input to triangulation function.
Per author of Modern Coding Theory Dr. Urbanke, the parity check matrix going into the encoding algorithm must be full rank, so I've added that constraint to the beginning of the GreedyUpperTriangulation function. Also, adding a step in the invMod2 function to cast the B matrix as integers - for some reason this was causing an error in the step to check that A*A^-1 is indeed the identity matrix.