Adding function getFullRankHmatrix.
It turns out that constructing parity check (H) matrices per Dr. Gallager's method will result it a few (actually the column weight-1) rows that are dependent. This new function takes a H matrix, and if it is not already full rank, performs an algorithm to find the dependent rows and delete them. The full rank H matrix is returned, but note that it will no longer be regular (if it was to begin with).