Updated sample encoding script using new functions.
This sample test creates a parity check matrix H of size 598x1200. It then encodes randomly chosen information bits of length k=602 into codewords of size 1200. Then a parity check on each codeword is done to ensure that it passes. This is repeated 500 times to create and check 500 codewords. They all passed when I ran this test.