Adding new function to read in alist files.
MacKay, Davey, and Lafferty share parity check matrices that they have developed in a file format they call alist. Adding a function readAlistFile in a new file LDPChelperFunctions.py. Also adding a test case which reads in a small alist file and compares it to the known H matrix, verifying that the readAlistFile function works.